extracted <DebugStateMessage/>
